Extract shortcuts to separate target
