+++ Blockzilla/UIComponents/UIConstants.swift
@@ -6,8 +6,6 @@ import UIKit
 
 struct UIConstants {
 
-    static let maximumNumberOfShortcuts = 4
-
     struct layout {
         static let browserToolbarDisabledOpacity: CGFloat = 0.4
         static let browserToolbarHeight: CGFloat = 44
@@ -101,12 +99,6 @@ struct UIConstants {
         static let introViewTextLabelInset: CGFloat = 24
         static let pageControlSpacing: CGFloat = 2
         static let toastMessageHeight: CGFloat = 48
-        static let shortcutViewWidth: CGFloat = 60
-        static let shortcutViewWidthIPad: CGFloat = 80
-        static let shortcutViewInnerDimension: CGFloat = 36
-        static let shortcutViewInnerDimensionIPad: CGFloat = 48
-        static let shortcutViewHeight: CGFloat = 84
-        static let shortcutViewHeightIPad: CGFloat = 100
         static let shortcutsContainerWidth: CGFloat = 326
         static let shortcutsContainerWidthIPad: CGFloat = 440
         static let shortcutsContainerOffset: CGFloat = 24
