+++ BlockzillaPackage/Sources/AppShortcuts/Shortcut.swift
@@ -4,11 +4,11 @@
 
 import UIKit
 
-struct Shortcut: Equatable, Codable, Hashable {
-    var url: URL
-    var name: String
+public struct Shortcut: Equatable, Codable, Hashable {
+    public var url: URL
+    public var name: String
     
-    init(url: URL, name: String = "") {
+    public init(url: URL, name: String = "") {
         self.url = url
         self.name = name.isEmpty ? Shortcut.defaultName(for: url) : name
     }
